Jetpack Connect: Move auth disclaimer to a separate component
